lib: renamed lib/stack-trace.js -> lib/trace.js
