+++ index.js
@@ -7,7 +7,7 @@ var EmitterDetails = require("./lib/emitter-details.js");
 var EventDetails = require("./lib/event-details.js");
 var HandlerDetails = require("./lib/handler-details.js");
 var helpers = require("./lib/helpers.js");
-var getCallSite = require("./lib/stack-trace.js");
+var getCallSite = require("./lib/trace.js");
 var debug = require("debug")("node-emitter-details");
 
 /**
