+++ lib/trace.js
